// src/models/InternalServerPublicProfile.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Exeq API
 * Remote on-demand browser infrastructure platform API
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@exeq.dev
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface InternalServerPublicProfile
 */
export interface InternalServerPublicProfile {
    /**
     * 
     * @type {string}
     * @memberof InternalServerPublicProfile
     */
    createdAt?: string;
    /**
     * 
     * @type {string}
     * @memberof InternalServerPublicProfile
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof InternalServerPublicProfile
     */
    name?: string;
}

/**
 * Check if a given object implements the InternalServerPublicProfile interface.
 */
export function instanceOfInternalServerPublicProfile(value: object): value is InternalServerPublicProfile {
    return true;
}

export function InternalServerPublicProfileFromJSON(json: any): InternalServerPublicProfile {
    return InternalServerPublicProfileFromJSONTyped(json, false);
}

export function InternalServerPublicProfileFromJSONTyped(json: any, ignoreDiscriminator: boolean): InternalServerPublicProfile {
    if (json == null) {
        return json;
    }
    return {
        
        'createdAt': json['created_at'] == null ? undefined : json['created_at'],
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
    };
}

export function InternalServerPublicProfileToJSON(json: any): InternalServerPublicProfile {
    return InternalServerPublicProfileToJSONTyped(json, false);
}

export function InternalServerPublicProfileToJSONTyped(value?: InternalServerPublicProfile | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_at': value['createdAt'],
        'id': value['id'],
        'name': value['name'],
    };
}