// src/models/InternalServerPublicExtendSessionRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Exeq API
 * Remote on-demand browser infrastructure platform API
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@exeq.dev
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface InternalServerPublicExtendSessionRequest
 */
export interface InternalServerPublicExtendSessionRequest {
    /**
     * 
     * @type {string}
     * @memberof InternalServerPublicExtendSessionRequest
     */
    duration?: string;
}

/**
 * Check if a given object implements the InternalServerPublicExtendSessionRequest interface.
 */
export function instanceOfInternalServerPublicExtendSessionRequest(value: object): value is InternalServerPublicExtendSessionRequest {
    return true;
}

export function InternalServerPublicExtendSessionRequestFromJSON(json: any): InternalServerPublicExtendSessionRequest {
    return InternalServerPublicExtendSessionRequestFromJSONTyped(json, false);
}

export function InternalServerPublicExtendSessionRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): InternalServerPublicExtendSessionRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'duration': json['duration'] == null ? undefined : json['duration'],
    };
}

export function InternalServerPublicExtendSessionRequestToJSON(json: any): InternalServerPublicExtendSessionRequest {
    return InternalServerPublicExtendSessionRequestToJSONTyped(json, false);
}

export function InternalServerPublicExtendSessionRequestToJSONTyped(value?: InternalServerPublicExtendSessionRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'duration': value['duration'],
    };
}